[Refactor][Xiaomeng] extract pipeline.ts.
