+++ frontend/src/models/pipeline.ts
@@ -1,0 +1,4 @@
+export enum PipelineTool {
+	BAMBOO = "BAMBOO",
+	JENKINS = "JENKINS",
+}
